% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_stratification.R
\name{age_strat_outcome_long}
\alias{age_strat_outcome_long}
\title{Age Stratify an Outcome Variable, long form}
\usage{
age_strat_outcome_long(outcome_var = "incidI", outcome_data, prob_data_list)
}
\arguments{
\item{outcome_var}{}

\item{outcome_data}{}

\item{prob_data_list}{}

\item{add}{}
}
\value{

}
\description{
Age Stratify an Outcome Variable, long form
}
