% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_stratification.R
\name{age_strat_outcome_wide}
\alias{age_strat_outcome_wide}
\title{#' Title
#'
#' @param data 
#' @param outcome_var 
#' @param prob_data 
#'
#' @return
#' @export
#'
#' @import dplyr
#' @importFrom purrr map
#' 
#' @examples
prob_fun_strat <- function(data,
                           outcome_var = "incidH",
                           prob_data){
    
    purrr:::map(data[[outcome_var]], rmultinom_t, n=1, prob = prob_data[[outcome_var]]) %>%
        dplyr::bind_rows() %>% setNames(paste0(outcome_var, "_", prob_data$age))
    
}
Age Stratify an Outcome Variable}
\usage{
age_strat_outcome_wide(
  outcome_var = "incidI",
  outcome_data,
  prob_data_list,
  add = TRUE
)
}
\arguments{
\item{outcome_var}{}

\item{outcome_data}{}

\item{prob_data_list}{}

\item{add}{}
}
\value{

}
\description{
#' Title
#'
#' @param data 
#' @param outcome_var 
#' @param prob_data 
#'
#' @return
#' @export
#'
#' @import dplyr
#' @importFrom purrr map
#' 
#' @examples
prob_fun_strat <- function(data,
                           outcome_var = "incidH",
                           prob_data){
    
    purrr:::map(data[[outcome_var]], rmultinom_t, n=1, prob = prob_data[[outcome_var]]) %>%
        dplyr::bind_rows() %>% setNames(paste0(outcome_var, "_", prob_data$age))
    
}
Age Stratify an Outcome Variable
}
