% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ageadjust.R
\name{convert_age_matrix_long}
\alias{convert_age_matrix_long}
\title{convert age matrix to long data}
\usage{
convert_age_matrix_long(age_matrix)
}
\arguments{
\item{age_matrix}{wide format matrix of age data, with row.name for geoid, and only columns for age groups. Column names need to be in the format of 10-19, 10_19, 10:19, or 10to19.}
}
\value{
long data.frame of age data
}
\description{
Fits a cubic spline across multiple studies to estimate the parameter of interest by specified age categories.
}
