% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{convert_wp_10yr}
\alias{convert_wp_10yr}
\title{Function to transform population data from a worldpop age groups to 10-year age groups}
\usage{
convert_wp_10yr(age_pop_data)
}
\arguments{
\item{age_pop_data}{data pulled from worldpop using `load_worldpop_age`}
}
\value{
long age population data by admin level 2, in 10-year age groups
}
\description{
Function to transform population data from a worldpop age groups to 10-year age groups
}
\references{
WorldPop (https://www.worldpop.org/geodata)
}
