% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ageadjust.R
\name{do_severity_sampling}
\alias{do_severity_sampling}
\title{Do age-specific parameter estimation}
\usage{
do_severity_sampling(
  param_to_est = "p_symp_inf",
  age_matrix,
  age_matrix_pct,
  raw_params,
  n_sims = 40,
  n_preds = 1000,
  age_grps = c(seq(0, 80, by = 10), 100),
  study_wt = "none"
)
}
\arguments{
\item{param_to_est}{parameter of interest}

\item{age_matrix}{wide format of age population}

\item{age_matrix_pct}{wide format of age population proportions}

\item{raw_params}{raw probabilities}

\item{n_sims}{numeric vector, the cutoff values for each age group}

\item{n_preds}{numeric, number of predictions for the gam model to make}

\item{age_grps}{age groups included in the estimation}

\item{study_wt}{study weight}
}
\value{
list
\itemize{
  \item \code{geoid_est}
  \item \code{sim_med} median simulation, to use for seeding
}
}
\description{
runs the full age estimation process. requires clean, setup age data setup using setup_age_data
}
