% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ageadjust_severe.R
\name{est_age_param}
\alias{est_age_param}
\title{Age-specific parameter estimation}
\usage{
est_age_param(
  expanded_dat,
  param_to_est = "p_symp_inf",
  age_cats = c(seq(0, 80, by = 10), 100),
  n_preds,
  study_wt = "none"
)
}
\arguments{
\item{expanded_dat}{data.frame with following columns:
\itemize{
  \item \code{param} name of the parameter to estimate
  \item \code{study} numeric, representing the study from which the observation came
  \item \code{age} numeric, age of the observed individual
  \item \code{x} numeric, binary 0 or 1 for whether outcome was observed
  }}

\item{param_to_est}{character, one of the parameter names in expanded_dat$param}

\item{age_cats}{numeric vector, the cutoff values for each age group}

\item{n_preds}{numeric, number of predictions for the gam model to make}

\item{study_wt}{character string, how to weight the random effects for the studies after the model is fit, there are multiple options:
\itemize{
  \item "none" (default) fits the random effects, but doesn't use them in the predictions of age-specific conditional probabilities
  \item "n" weights each random effect by the number of observations in each study
  \item "root_n" weights each random effect by the square root of the number of observations
  \item "equal" weights each random effect equally
}}
}
\value{
list with 3 items
\itemize{
  \item \code{pred_mtx} matrix, rows are age categories, columns are simulations, values are estimates of the conditional probabilities
  \item \code{pred_sum} tbl, with the median, lower and upper bounds of a 95\% CI for each age category
  \item \code{param_to_est} character_string, same as in Arguments
}
}
\description{
Fits a cubic spline across multiple studies to estimate the parameter of interest by specified age categories.
}
\examples{
## dummy data
expanded_dat <- tibble(param="p_vent_icu",
                       study=c(rep(1,400), rep(2,600)),
                       age=runif(1000,0,100),
                       x=c(rbinom(400, 1, 0.3), rbinom(600, 1, 0.45)))

## run function
p_vent <- est_age_param(expanded_dat=expanded_dat,
                             param_to_est="p_vent_icu",
                             n_preds=1000)

## view summary
p_vent$pred_sum
}
