% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ageadjust_severe.R
\name{est_geoid_rrs}
\alias{est_geoid_rrs}
\title{Estimate relative rates for GEOIDs}
\usage{
est_geoid_rrs(
  pred_mtx,
  param_to_est,
  geoid_age_mtx = US_age_geoid_pct,
  geoid_pops = US_age_geoid_pop
)
}
\arguments{
\item{pred_mtx}{matrix of predicted parameter values by age category and simulation from est_age_param()$pred_mtx}

\item{param_to_est}{character string to name output column}

\item{geoid_age_mtx}{matrix of proportion of population by age category and GEOID (e.g. data(US_age_geoid_pct))}

\item{geoid_pops}{matrix of total population by age category and GEOID (e.g. data(US_age_geoid_pop))}
}
\value{
tbl, with relative rates by GEOID and the overall estimate across all GEOIDs
}
\description{
Estimate relative rates for GEOIDs
}
