% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ageadjust_severe.R
\name{expand_age_data}
\alias{expand_age_data}
\title{Expanded age category data}
\usage{
expand_age_data(age_cat_data, param_to_est = "p_symp_inf")
}
\arguments{
\item{age_cat_data}{data.frame}

\item{param_to_est}{character string of parameter of interest}
}
\value{
data.frame with following columns:
\itemize{
  \item \code{param} name of the parameter to estimate
  \item \code{study} integer, representing the study from which the observation came
  \item \code{age} numeric, age of the observed individual
  \item \code{x} numeric, binary 0 or 1 for whether outcome was observed
  }
}
\description{
draw ages from categories for better model fitting
}
\examples{
## load param data
data("raw_age_estimates")
## filter to only those for pipeline
raw_params <- filter(raw_age_estimates, USE_pipeline==T)
## expand data for "p_symp_inf"
expand_age_data(raw_params, "p_symp_inf")
}
