% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{format_age_county_pop}
\alias{format_age_county_pop}
\title{Function to format age-specific county population into desired bins}
\usage{
format_age_county_pop(
  age_county_pop,
  age_lower = seq(0, 80, 10),
  age_upper = c(seq(9, 79, 10), 99),
  agecat_labels = NULL
)
}
\arguments{
\item{age_county_pop}{ACS data from load_age_county_pop}

\item{age_lower}{lower limit of desired age categories}

\item{age_upper}{upper limit of desired age categories}

\item{agecat_labels}{if NULL, will be paste0(age_lower, "_", age_upper)}
}
\value{
a data frame with columns
         - GEOID
         - cat_l (index of age category)
         - estimate
         - NAME
         - age_l
         - age_r
         - agecat
}
\description{
Function to format age-specific county population into desired bins
}
