% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_stratification.R
\name{get_age_strat_outcomes}
\alias{get_age_strat_outcomes}
\title{Title}
\usage{
get_age_strat_outcomes(
  outcomes = c("incidI", "incidMild", "incidH", "incidICU", "incidVent", "incidD"),
  outcome_data,
  prob_data,
  wide = FALSE
)
}
\arguments{
\item{outcomes}{}

\item{outcome_data}{}

\item{prob_data}{}
}
\value{

}
\description{
Title
}
