% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ageadjust.R
\name{get_ageadjustments}
\alias{get_ageadjustments}
\title{Do age-specific parameter estimation}
\usage{
get_ageadjustments(
  age_data,
  cores = 4,
  n_sims = 40,
  n_preds = 1000,
  age_grps = c(seq(0, 80, by = 10), 100),
  googlesheet_access = FALSE,
  output_dir = "output",
  pop_name = NULL
)
}
\arguments{
\item{age_data}{long form age data with 10-year age groups and with columns: [geoid, age, pop, prop]}

\item{cores}{integer, number of cores to use for parallel processing}

\item{n_sims}{numeric vector, the cutoff values for each age group}

\item{n_preds}{numeric, number of predictions for the gam model to make}

\item{age_grps}{age groups included in the estimation}

\item{googlesheet_access}{whether the user has access to the raw data on google docs; default is FALSE}

\item{output_dir}{where the user wants the output data to be saved}
}
\value{
list
}
\description{
runs the full age estimation process. requires clean, setup age data setup using setup_age_data
}
