% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ageadjust.R
\name{get_allcountries_age_pop}
\alias{get_allcountries_age_pop}
\title{Get population distribution from WPP and aggregate it to 10 year age groups
- this is set up to use population estimates from the World Populaiton Prospects data}
\usage{
get_allcountries_age_pop()
}
\arguments{
\item{country}{country of interest}

\item{ISO3}{Optional; ISO3 code for country of interest.}
}
\value{
10-year age group populations for country of interest, long form
}
\description{
Get population distribution from WPP and aggregate it to 10 year age groups
- this is set up to use population estimates from the World Populaiton Prospects data
}
