% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_stratification.R
\name{get_p_age_given_outcome}
\alias{get_p_age_given_outcome}
\title{Multiply Pr(age|var) * Pr(age) for each geoid
instead of inputting p_hosp to normalize to 1, we just normalize to 1}
\usage{
get_p_age_given_outcome(
  p_var = "p_hosp_inf",
  p_age_long,
  params_age,
  age_groups
)
}
\arguments{
\item{p_var}{}

\item{p_age_long}{}

\item{params_age}{}
}
\value{

}
\description{
Multiply Pr(age|var) * Pr(age) for each geoid
instead of inputting p_hosp to normalize to 1, we just normalize to 1
}
