% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_stratification.R
\name{get_prob_ages_all}
\alias{get_prob_ages_all}
\title{Title}
\usage{
get_prob_ages_all(p_age_long, params_age)
}
\arguments{
\item{p_age_long}{}

\item{params_age}{}
}
\value{

}
\description{
Title
}
