% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{load_age_county_pop}
\alias{load_age_county_pop}
\title{Function to load age-specific population by county for states + year of interest}
\usage{
load_age_county_pop(census_api_key, modeled_states, census_year)
}
\arguments{
\item{census_api_key}{census API key required for tidycensus}

\item{modeled_states}{states for which to pull data}

\item{census_year}{year for which to generate data}
}
\value{
a data frame with columns
         - GEOID
         - agecat
         - estimate
         - NAME
         - age_l
         - age_r
}
\description{
Function to load age-specific population by county for states + year of interest
}
