% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{load_worldpop}
\alias{load_worldpop}
\title{Function to get population data from a worldpop .tif and geounit shapefile
This function extracts raster values for overlayed polygons}
\usage{
load_worldpop(wp_file, shp)
}
\arguments{
\item{wp_file}{WorldPop population file path for the country of interest}

\item{shp}{Shapefile file path for shapefile of geunits (i.e., admin2)}
}
\value{

}
\description{
Function to get population data from a worldpop .tif and geounit shapefile
This function extracts raster values for overlayed polygons
}
\references{
WorldPop (https://www.worldpop.org/geodata)
}
