% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{load_worldpop_age}
\alias{load_worldpop_age}
\title{Function to get population data from a worldpop geotiff and geounit shapefile
This function extracts raster values for overlayed polygons}
\usage{
load_worldpop_age(
  shp,
  country = "BGD",
  year = "2020",
  save_dir = "raw_data",
  cores = 4,
  loc_var = NA
)
}
\arguments{
\item{shp}{Shapefile file path for shapefile of geunits (i.e., admin2)}

\item{country}{ISO3 of country of interest}

\item{year}{Year of population data (2000 to 2020)}

\item{save_dir}{directory where to save geotiff files}

\item{cores}{number of cores to parallelize over}

\item{loc_var}{name of location name or id variable, if want to reduce data to this and population data}
}
\value{
long age population data by admin level 2
}
\description{
Function to get population data from a worldpop geotiff and geounit shapefile
This function extracts raster values for overlayed polygons
}
\references{
WorldPop (https://www.worldpop.org/geodata)
}
