% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{make_age_matrix}
\alias{make_age_matrix}
\title{Function to make population data into matrix for multiplication}
\usage{
make_age_matrix(age_county_pop_formatted)
}
\arguments{
\item{age_county_pop_formatted}{formatted ACS data from format_age_county_pop}
}
\value{
a listof GEOID vector, agecat labels, and p_age matrix
}
\description{
Function to make population data into matrix for multiplication
}
