% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_stratification.R
\name{prob_fun}
\alias{prob_fun}
\title{Title}
\usage{
prob_fun(data, outcome_var = "incidH", prob_data, incl_outcome = TRUE)
}
\arguments{
\item{data}{}

\item{outcome_var}{}

\item{prob_data}{}
}
\value{

}
\description{
Title
}
