% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-descriptions.R
\docType{data}
\name{raw_age_estimates}
\alias{raw_age_estimates}
\title{Raw age-specific estimates}
\format{tbl with the following columns
\itemize{
  \item \code{param} character string of a conditional probability with the format p_x_y that is equivalent to P(X | Y)
  \item \code{USE_pipeline} logical, is the estimate used in the model pipeline?
  \item \code{ageL} numeric, lower bound of age category
  \item \code{ageR} numeric, upper bound of age category
  \item \code{X} integer, numerator value of conditional probability (e.g., the number of symptomatic cases in P(symptomatic | infection))
  \item \code{N} integer, denominator value of conditional probability (e.g., the number of infections in P(symptomatic | infection))
  \item \code{value} numeric, the proportion estimate for the conditional probability and age category.Is equal to X/N when both are known.
  \item \code{valueL} numeric, lower bound of the conditional probability
  \item \code{valueR} numeric, upper bound of the conditional probability
  \item \code{desc} character string, description of the observation
  \item \code{data_date} date, when was the data collected
  \item \code{publish_date} date, when was the source published
  \item \code{source} character string, where is the data from
}}
\usage{
data(raw_age_estimates)
}
\description{
A data frame with age-specific parameter estimates from a variety of sources
}
