% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ageadjust.R
\name{setup_age_data}
\alias{setup_age_data}
\title{Setup age data for the parameter estimation}
\usage{
setup_age_data(age_data)
}
\arguments{
\item{age_data}{long format age data in 10 year age groups, with columns: [geoid, age, pop, prob]}
}
\value{
list with 3 items
\itemize{
  \item \code{age_matrix} matrix, rows are age categories, columns are simulations, values are estimates of the conditional probabilities
  \item \code{age_matrix_pct} tbl, with the median, lower and upper bounds of a 95\% CI for each age category
  \item \code{age_data_long} character_string, same as in Arguments
}
}
\description{
Setup age data for the parameter estimation
}
