% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-descriptions.R
\docType{data}
\name{US_age_geoid_pct}
\alias{US_age_geoid_pct}
\title{US population by age category and GEOID}
\format{
A matrix where rows are GEOIDs and columns are age groups
}
\usage{
data(US_age_geoid_pct)
}
\description{
A matrix containing the proportion of the population by 10 year age
categories in the US
}
