% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-descriptions.R
\docType{data}
\name{US_geoid_params}
\alias{US_geoid_params}
\title{Age-adjusted parameters for each US GEOID}
\format{tbl, where rows are GEOIDs and columns are parameters}
\usage{
data(US_geoid_params)
}
\description{
A data frame with the median conditional probabilities for each parameter for each GEOID
}
