% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ageadjust_severe.R
\name{est_age_param}
\alias{est_age_param}
\title{Age-specific parameter estimation}
\usage{
est_age_param(
  age_cat_data,
  param_to_est = "p_symp_inf",
  age_cats = c(seq(0, 80, by = 10), 100),
  n_preds,
  study_wt = "none"
)
}
\arguments{
\item{param_to_est}{character, one of the parameter names in age_cat_data$param}

\item{age_cats}{numeric vector, the cutoff values for each age group}

\item{n_preds}{numeric, number of predictions for the gam model to make}

\item{study_wt}{character string, how to weight the random effects for the studies after the model is fit, there are multiple options:
\itemize{
  \item "none" (default) fits the random effects, but doesn't use them in the predictions of age-specific conditional probabilities
  \item "n" weights each random effect by the number of observations in each study
  \item "root_n" weights each random effect by the square root of the number of observations
  \item "equal" weights each random effect equally
}}
}
\value{
list with 3 items
\itemize{
  \item \code{pred_mtx} matrix, rows are age categories, columns are simulations, values are estimates of the conditional probabilities
  \item \code{pred_sum} tbl, with the median, lower and upper bounds of a 95\% CI for each age category
  \item \code{param_to_est} character_string, same as in Arguments
}
}
\description{
Fits a cubic spline across multiple studies to estimate the parameter of interest by specified age categories.
}
\examples{
## load param data
data("raw_age_estimates")
## filter to only those for pipeline
raw_params <- filter(raw_age_estimates, USE_pipeline==T)

## run function
p_vent <- est_age_param(age_cat_data=raw_params,
                        param_to_est="p_vent_icu",
                        n_preds=1000)

## view summary
p_vent$pred_sum
}
