% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ageadjust_severe.R
\name{est_geoid_params}
\alias{est_geoid_params}
\title{Find parameter values for GEOIDs}
\usage{
est_geoid_params(age_params, geoid_age_mtx = US_age_geoid_pct)
}
\arguments{
\item{age_params}{list, output of est_age_param()}

\item{geoid_age_mtx}{matrix of population by age and GEOID (e.g. data(US_age_geoid_pct))}
}
\value{
tbl, median probabilities for the given parameter by GEOID
}
\description{
Find parameter values for GEOIDs
}
